Required Style Headers
Issue #50
Check that all required headers are included in style.css.
Requires WordPress_AbstractThemeSniff class.
For Discussion
==========
If of the required header labels are not included, that is if the line is not there at all, then this check will not display. This could be avoided by setting stackPtr to 0 with the addWarning message.