:rocket: Fix travis build: load PHPCS native bootstrap if available and test against PHP 7.2
The upstream PR to support PHPUnit 6.x has (finally) been merged and will be released in PHPCS 3.1.

Currently builds against PHPCS `master` are failing because of this as the final version of the PHPCS PHPUnit 6.x support PR includes a test bootstrap in PHPCS which is not being loaded as we use our own bootstrap.

Easily solved though, by loading the upstream bootstrap - if the file exists - from our own PHPCS 3 bootstrap file. That way our unit testing will be compatible with all supported PHPCS 3.x versions.

The `beStrictAboutTestsThatDoNotTestAnything="false"` addition to the `phpunit.xml.dist` file is needed to prevent PHPUnit from reporting there are no tests in our test files (as the tests are in the upstream test suite and our "tests" are basically only data providers).

The changes made in 873 should, for now, not be reverted as PHPCS < 3.1 is not compatible with PHPUnit 6.x and accounting for all the different situation with an `if` statement in the build script would get unnecessarily complicated. See the table below for more details.

Once PHPCS 2.x support is dropped and the minimum supported PHPCS version has gone up to 3.1, we should be able to revert the changes made in 873 without issues (except for HHVM).

Upstream references:
* squizlabs/PHP_CodeSniffer 1383
* squizlabs/PHP_CodeSniffer 1384

Additionally:
* While looking at the travis builds to fix this I noticed that `nightly` has gone up to PHP `7.3.0-dev`, so I've added PHP 7.2 to the build matrix.