PHPCS 3.x compat: Ruleset: Make sure the complete sniff is included, not just an errorcode
In PHPCS 2.x when a sniff was referenced by errorcode, this automatically included the whole sniff.
As of PHPCS 3.x it is possible to selectively include just and only a specific errorcode.

This allows for cross-version compat and ensures that if/when new errorcodes are added to the sniffs, they will be included as well.