DeprecatedWPConstants: Remove sniff in favour of WPCS sniff
WPCS 0.14.0 introduced a `WP.DiscouragedConstants` sniff which covers all the constants which were covered by the `DeprecatedWPConstants` sniff and does so better than the original sniff.

Removed the `DeprecatedWPConstants` sniff in favour of the new upstream sniff and changing the error level from `warning` to `error` from the ruleset maintains the existing behaviour.

Ref: WordPress-Coding-Standards/WordPress-Coding-Standards 1089