GH issue template: remove WPCS version in favour of TRTCS one
The backmerge from WPCS added a second GH issue template to the `.github` directory which apparently takes precedence over the TRTCS version which was already there.

Removing the WPCS version fixes this.

----

N.B.: This would (will) need to be done each time after a WPCS back-merge, but as the repo is probably going to be a stand-alone standard, this will soon no longer be an issue.