Add check for whitespace before PHP open tag
This adds an upstream sniff which can check for whitespace before a PHP open tag.
As the sniff does more than just that, all other error codes from the sniff are excluded.

The check takes inline HTML into account and will not throw an error is the PHP open tag is not the first non-whitespace token in the file.

Notes:
* Changed the error message to make it very clear why this check is here.
* The upstream sniff does not handle unicode whitespace well. This is being discussed in an issue I opened upstream and will probably/hopefully be fixed soon.

Fixes 148