:twisted_rightwards_arrows: Recategorize/TestCase files: adjust `phpcs:set` annotations
Important note: as the TGMPA version sniff throws errors on line 1/first token, the first token can *not* be the annotation as PHPCS 3.2+ ignores errors/warnings on lines which _only_ contain annotations. So to make the annotations work for these unit tests, they have to be on the second line, i.e. not the first token.