:scissors: Add WPThemeReview native autoload file
As several sniffs in the WPThemeReview standard extend (abstract) sniffs in WPCS, the WPCS autoload file needs to be loaded.

If WPThemeReview would load any of the WPCS native rulesets in its entirety, this would be done automatically, however that is not the case.

For end-users the only supported installation method for WPThemeReview is installing via Composer.
However, developers for WPThemeReview may have a more complex setup using git clones of PHPCS and WPCS.

So, the autoload file has been setup to support both.
Developers for WPThemeReview who don't use Composer, can either set a `WPCS_DIR` environment variable or add a `.pathtowpcs` file to the project root containing the path to the WPCS install on their system.

More information about this will be added to the `CONTRIBUTING.MD` file.