Update for PHPCompatibility 9.0.0
[PHPCompatibility 9.0.0](https://github.com/PHPCompatibility/PHPCompatibility/releases/tag/9.0.0) has just been released and renames all sniffs.
[PHPCompatibilityWP 2.0.0](https://github.com/PHPCompatibility/PHPCompatibilityWP/releases/tag/2.0.0) has been released alongside it.

The upgrade for this repo is simple as TRTCS currently doesn't use PHPCS inline annotations which reference PHPCompatibility sniffs, nor uses custom excludes.

As the PHPCompatibilityWP ruleset as of version 2.0 has an additional dependency, it is however opportune to let the DealerDirect plugin sort out the installed paths, at least for devs.

For end-users, the DealerDirect plugin is advised anyway, so hopefully people will listen to that advice ;-)

* Updated the version restrictions in `composer.json`.
* Removed the `install-standards-dev`, `post-install-cmd` and `post-update-cmd` scripts as that will now be sorted out by the DealerDirect plugin.
* Updated the Readme references.