Travis: test against high/low WPCS
As there are quite some sniffs which now extend WPCS sniffs, it would be prudent to test builds against a mix of PHPCS and WPCS versions.

With the changes made in this PR, all sniffs will be tested against a combination of the highest PHPCS+WPCS versions as well as the lowest PHPCS+WPCS versions on most PHP versions.

To also make sure that the sniffs work correctly with a different combi, the PHP 7.2 build will test with high PHPCS vs low WPCS and with low PHPCS vs high WPCS.