Trying to utilise GH actions for running checks instead of travis (#261)
* Add the sniff stage

* Remove travis file

No need to clutter checks

* Setup xml lint

* Remove XMLLINT_INDENT export

* Add empty test.yaml file

* Fix the diff?

* Trying to fix the diff

* Try to run on xenial

* Add lint check and composer validation to workflow

* Trying to add test stage

* Try to exclude certain versions

Only allow master phpcs and develop wpcs on one PHP version, and that should be allowed to fail

* Refine exclusions

We only need master - master and lowest - lowest for every PHP version, not all 4 possible combinations.

* PHP exclusions cannot be in the list apparently

* Adding php 5.4 and 5.5

* Fix typo

* Add nightly to build

* Fix the sniff workflow

* Update suggestions for dealerdirect plugin

So that it works with the composer v2

* Remove travis reference

* Fix the minor grammar issues in the Contributing file

Also change the theme review team repo to correct repo name. It's themes team now.

* Fix the sniff stage accordding to the PR suggestions

Also, some ideas from PHPCompatibility PR https://github.com/PHPCompatibility/PHPCompatibility/pull/1260 were implemented as well (composer dependencies and xml violations).

* Update test workflow with the suggestions from the PR

* Trying to fix the invalid yaml issue

* Change quotation marks in the condition

* Specify the PHP 8 branch that can, and will, fail

* Add continue-on-error directive

* Fix the badges and the repo name

* GH Actions: remove the ignores for markdown files

... as it would prevent PRs with only markdown changes from being merged due to required statuses.

* GH Actions: allow for manually triggering a workflow

Triggering a workflow for a branch manually is not supported by default in GH Actions, but has to be explicitly allowed.

Ref: https://github.blog/changelog/2020-07-06-github-actions-manual-triggers-with-workflow_dispatch/

* GH Actions: remove GH token from setup PHP step

... as it shouldn't be needed.

* GH Actions: add some inline documentation for some steps

* GH Actions: add lint workflow

As things were, the code was only ever linted against PHP 7.4, while in the original Travis workflow, the code was linted against every single supported PHP version.

This adds a hybrid workflow in which the code is linted against the high/low of each PHP major within the range of supported PHP versions.

* GH Actions: use Parallel-Lint for PHP linting

Not a necessity, but allows for:
1. Faster linting due to the parallel option.
2. The ability to display linting errors in the PR via the cs2pr tooling.

* GH Actions: fix up the matrix

The minimum supported PHPCS version according to the `composer.json` file, is `3.3.1`, so that should be the minimum version to test against.
Similarly, the minimum supported WPCS version according to the `composer.json` file is `2.2.0`.

For PHP 7.4 and 8.0, the builds have to be specified explicitly as PHP 7.4 only has PHPCS runtime support as of PHPCS 3.5.0 and PHP 8.0, as of PHPCS 3.5.7.

As the matrix builds only test high PHPCS with high WPCS and low PHPCS with low WPCS, for PHP 7.2, in the Travis config, explicit builds were specified to reverse that combination, i.e. high PHPCS with low WPCS and low PHPCS with high WPCS.
This has now been implemented here as well.

Includes adding some inline documentation to the matrix.

Note: I've not added a build against PHP 8.1/nightly as the PHPUnit version we are forced to use via PHPCS (PHPUnit 7.x) is not compatible with PHP 8.1, so the build wouldn't pass no matter what.

* GH actions: set the PHP ini values

* GH actions: remove duplication of setting of PHPCS/WPCS versions

* GH actions: handle Composer install on PHP 8

* Readme: fix up the badges

... based on latest example code generate on the Actions page.

* Remove extra quotes and reduced spacing from 4 to 2 spaces

Co-authored-by: jrfnl <jrfnl@users.noreply.github.com>